Update README.md
The safety of motorcycles and all the included variables has been a concern for years. Meanwhile vehicles have dramatically increased their safety ratings by integrating technological advances such as using sensors that help alert motorists and stay focused. Continual advances in technology have added numerous safety systems. We intend to make a device that will use similar technology on a motorcycle in order to garner the same outcomes for motorcyclists. The Defensive Alert Safety Unit (DASU) will be our contribution to keep motorcycles as advanced as motor vehicles.

DASU is an auxiliary system intended to keep motorcyclist safe and alert to the dangers around them, while keeping the motorists aware of the motorcyclist’s presence on the road. Utilizing project management practices and engineering design processes, we intend to build a fully functioning prototype.
 
DASU will use a microcontroller and sensors to maintain a safe perimeter around the motorcycle. Whenever this perimeter is breached, DASU will alert the motorcyclist of the dangers with the hopes that the motorcyclist drive defensively. By allowing the motorcyclist to take action during the early signs of danger, DASU will greatly reduce the risks faced by motorcyclists. In the event that these parameters are not enough and danger is imminent, DASU will execute a command that will sound a siren or horn and flash the motorcycle lights. This course of action will make the motorcycle more visible and catch the attention of the motorist.
 
The Defensive Alert Safety Unit is designed to decrease the number of motorcycle accidents and has the potential to save lives. The essential goal is to get a fully functioning system that meet the demands of safety regulations, and then through further research and development, place supplementary systems on the DASU to increase its value and efficacy. We hope to increase motorcyclist’s safety awareness thus allowing them to focus on the pleasures that attracts riders to motorcycles.